% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_gun_manufacturing_1860}
\alias{us_gun_manufacturing_1860}
\title{Gun Manufacturers by Region Dataset}
\format{
A data frame with the following columns:
\describe{
\item{region}{(character) The name of the region where gun manufacturers are located.}
\item{no_gun_manufacturers}{(integer) The number of gun manufacturing companies or facilities located within the region.}
\item{value_of_product}{(numeric) The total value of products manufactured by gun manufacturers in the region. This value is in dollars.}
}
}
\source{
Manufactures of The United States in 1860; Compiled From the Original Returns of the Eighth Census under the direction of
the Secretary of the Interior, Bureau of Census Library, pp. cxc.
}
\usage{
data(us_gun_manufacturing_1860)
}
\description{
This dataset provides information about gun manufacturers by region, including the number of gun manufacturers and the total value of products manufactured.
}
\details{
This dataset is intended for analyzing the distribution of gun manufacturing companies and the corresponding economic value of their products across different regions. It can be used for market analysis, economic studies, and regional assessments of industrial activity.
}
\examples{
# Example of how to load and use the dataset
data(us_gun_manufacturing_1860)
head(us_gun_manufacturing_1860)

}
\keyword{datasets}
