% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_railroad_miles_1860}
\alias{us_railroad_miles_1860}
\title{U.S. Railroad Miles in 1860 by Region}
\format{
A data frame with the following columns:
\describe{
\item{region}{(character) The name of the region (state or other geographical area) in the United States.}
\item{railroad_miles_1860}{(numeric) The total miles of railroad tracks in the region as of 1860.}
}
}
\source{
Data was sourced from page 185 of the book "One Hundred Years of American Commerce" by Chauncey M. Depew, available at \url{https://babel.hathitrust.org/cgi/pt?id=ien.35556015719511&seq=185}.
}
\usage{
data(us_railroad_miles_1860)
}
\description{
This dataset provides information on the total miles of railroad tracks by region in the United States in 1860.
}
\details{
This dataset contains historical information about the expansion of the railroad system in the United States as of 1860. It can be used to analyze the distribution of railroad infrastructure across different regions, providing insights into the early industrial and transportation development of the country.
}
\examples{
# Example of how to load and use the dataset
data(us_railroad_miles_1860)
head(us_railroad_miles_1860)

}
\keyword{datasets}
