% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_census_establishments_1860}
\alias{us_census_establishments_1860}
\title{Industrial and Economic Indicators by State and Year}
\format{
A data frame with X rows and 8 columns:
\describe{
\item{\code{year}}{Year in which the data was collected (integer).}
\item{\code{state}}{State or region (character).}
\item{\code{no_of_establishments}}{Number of industrial establishments (integer).}
\item{\code{total_capital_invested}}{Total capital invested in the establishments (numeric).}
\item{\code{annual_value_of_products}}{Annual value of products produced by the establishments (numeric).}
}
}
\source{
Manufactures of The United States in 1860; Compiled From the Original Returns of the Eighth Census under the direction of
the Secretary of the Interior, Bureau of Census Library, pp. 729-730.
}
\usage{
data(us_census_establishments_1860)
}
\description{
This dataset contains industrial and economic data for U.S. states, including the number of establishments (manufactures),
total capital investment, and the value of annual products.
}
\examples{
# Load the dataset
data(us_census_establishments_1860)

# View the first few rows
head(us_census_establishments_1860)
}
\keyword{datasets}
